requirements.txt
streamlit
whisper
moviepy
google-generativeai
python-dotenv
ffmpeg-python
textstat